ai_startup_feature_collector/pyproject.toml
[project]
name = "ai-startup-feature-collector"
version = "0.1.0"
description = "End-to-end pipeline for collecting AI startup intelligence features"
authors = [{name = "OpenAI Assistant"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.0"
]

[project.optional-dependencies]
serp = ["google-search-results>=2.4.0"]
aio = ["aiohttp>=3.9.0"]
rag = ["langchain>=0.1.0", "chromadb>=0.4.0"]

[project.urls]
"Homepage" = "https://github.com/example/ai-startup-feature-collector"
"Documentation" = "https://github.com/example/ai-startup-feature-collector"
